+++ time_complexity/biconnected.hpp
@@ -12,16 +12,16 @@ class FaceTriangulation;
 class biconnected
 {
 public:
-    vector<vector<int>> faces;
-    unordered_set<pair<int, int>, PairHash> present;
-    vector<vector<pair<int, int>>> allTriangulations;
+    vector<vector<long long>> faces;
+    unordered_set<pair<long long, long long>, PairHash> present;
+    vector<vector<pair<long long, long long>>> allTriangulations;
     vector<FaceTriangulation *> faceTriangulations;
     u128 totalTriangulations = 0;
 
-    biconnected(vector<vector<int>> &faces)
+    biconnected(vector<vector<long long>> &faces)
     {
         this->faces = faces;
-        present = unordered_set<pair<int, int>, PairHash>();
+        present = unordered_set<pair<long long, long long>, PairHash>();
         initiatePresent();
         faceTriangulations = vector<FaceTriangulation *>(faces.size(), nullptr);
     }
@@ -32,7 +32,7 @@ public:
     {
         for (auto face : faces)
         {
-            for (int i = 0; i < face.size() - 1; i++)
+            for (long long i = 0; i < face.size() - 1; i++)
             {
                 present.insert(make_pair(min(face[i], face[i + 1]), max(face[i], face[i + 1])));
             }
@@ -40,7 +40,7 @@ public:
         }
     }
     void getAllTriangulations();
-    void output(int serial);
+    void output(long long serial);
     void addTriangulation();
     void sortTriangulations()
     {
@@ -80,7 +80,7 @@ inline void biconnected::getAllTriangulations()
     // printAllTriangulations();
 }
 
-inline void biconnected::output(int serial)
+inline void biconnected::output(long long serial)
 {
     if (serial == faces.size() - 1)
     {
@@ -97,10 +97,10 @@ inline void biconnected::output(int serial)
 
 // inline void biconnected::addTriangulation()
 // {
-//     vector<pair<int, int>> currentTriangulations;
+//     vector<pair<long long, long long>> currentTriangulations;
 //     for (auto a : faceTriangulations)
 //     {
-//         vector<pair<int, int>> currentTriangulation;
+//         vector<pair<long long, long long>> currentTriangulation;
 //         for (auto &chord : a->chords)
 //         {
 //             currentTriangulation.push_back(a->getPair(chord));
